+++ RigidBody.cs
@@ -27,7 +27,7 @@ class Rigidbody : Component
 
     public void GravityHandler()
     {
-        MyVector2 displacement = new(position.X, position.Y + Gravity * 5f);
+        MyVector2 displacement = new(position.X, position.Y + (int)Gravity * 5);
         gameobject.transform.Position = new MyVector2
             (gameobject.transform.Position.X + displacement.X, gameobject.transform.Position.Y + displacement.Y);
     }
